+++ x/staging/posts/simulation/params.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 
-	"github.com/desmos-labs/desmos/x/staging/posts/types"
+	"github.com/desmos-labs/desmos/v2/x/staging/posts/types"
 )
 
 func ParamChanges(r *rand.Rand) []simtypes.ParamChange {